Modified Camera and Road Spawning
Modified the camera to ensure that it wasn't showing to far in the distance, this involved changing the rotation and position of where the camera sat, this is needed to ensur not to many items road sections need to be rendered on the screen at once and that the road sections can be instantiated as the player moves towards them seemlessly. As a result the game manager was modifid to perform a less accurate check on the player position before spawning road sections.